refactor: simplify api dir structure
